Update README.md
This project involves visualization of the covid live data available through oxford covid live tracker. The main objective of this project is to visualizes the covid data of four countries United kingdom, Sweden, South korea and Georgia. These four countries have adopted four unique lock-down measures for covid-19. The confirmed cases, confirmed deaths and Infection rate of this four countries are analyzed and visualizations are provided appropriately. The key idea behind this project is to make use of the visualized data and consider their respective populations to decide the best lock-down measure to be adopted by other countries in the face of a second lock-down.